Project Modified for Pipex project
This project was modified for help me to do the next project.

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: lumorale <lumorale@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/28 13:53:45 by lumorales         #+#    #+#             */
-/*   Updated: 2022/12/09 13:15:00 by lumorale         ###   ########.fr       */
+/*   Updated: 2023/02/20 17:19:17 by lumorale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	counter;
